renamed files after discussion on APIs
